test(web): split UI e2e suite into focused spec files
